backend/requirements.txt
# Core Django
Django>=5.0,<5.1
djangorestframework>=3.14.0
djangorestframework-gis>=1.0
django-cors-headers>=4.3.0
django-filter>=23.5
django-environ>=0.11.2

# Database
psycopg2-binary>=2.9.9

# GeoDjango / GIS (GeoDjango is part of Django)
shapely>=2.0.2
pyproj>=3.6.1

# Celery
celery>=5.3.0
redis>=5.0.0
django-celery-beat>=2.5.0

# SGP4 / Astronomy
skyfield>=1.46
sgp4>=2.23

# Keycloak / Auth
python-keycloak>=3.6.0
PyJWT>=2.8.0
cryptography>=41.0.0

# Data processing
pandas>=2.1.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# API / HTTP
requests>=2.31.0
httpx>=0.25.0

# Validation
pydantic>=2.5.0

# Utils
python-dateutil>=2.8.0
pytz>=2023.3

# Production
gunicorn>=21.2.0
whitenoise>=6.6.0

# Development
pytest>=7.4.0
pytest-django>=4.7.0
factory-boy>=3.3.0
black>=23.0.0
flake8>=6.0.0
